Changed some parse.h and parser.c stuff
trying to redefine our trees so that they follow the AST in the book (screenshot uploaded to info folder in this branch). So we have: assignment is a left side identifier and a right side that points to something, either another identifier, a literal, or a binary thingy. A binary has a character that is an operator, and a leftside and rightside that are literals or another binary, etc, Then inside parser.c the functions newAssignment, newBinary, etc would make sure that the structs are only used when they need to be and their left and right sides and such make sense. Also added global arrays of assignments and identifiers as we need to save all the structs we make. Plus we'll need them in the type checker I think.